+++ src/query/get_room_information/mod.rs
@@ -1,0 +1,2 @@
+//! Endpoint to query room information with a room alias.
+pub mod v1;